move getEnv out of utils
