+++ discovery-provider/plugins/network-monitoring/src/db/index.ts
@@ -2,7 +2,7 @@
 import { Sequelize } from 'sequelize'
 
 import { SequelizeStorage, Umzug } from 'umzug'
-import { getEnv } from '../utils'
+import { getEnv } from '../config'
 import create_foreign_connection from './migrations/create_foreign_connection'
 import create_tables from './migrations/create_tables'
