+++ discovery-provider/plugins/network-monitoring/src/metrics/index.ts
@@ -11,7 +11,7 @@ import {
   userCountGauge,
   usersWithAllFoundationNodeReplicaSetGauge,
 } from "../prometheus";
-import { getEnv } from "../utils";
+import { getEnv } from "../config";
 import {
   getPrimaryUserCount,
   getAllUserCount,
